Move '_require_connection' def into related '_implicit_environ' module.
