+++ gcloud/datastore/test__implicit_environ.py
@@ -299,6 +299,58 @@ class Test_lazy_loading(unittest2.TestCase):
             'connection' in _implicit_environ._DEFAULTS.__dict__)
 
 
+class Test__require_connection(unittest2.TestCase):
+
+    _MARKER = object()
+
+    def _callFUT(self, passed=_MARKER):
+        from gcloud.datastore._implicit_environ import _require_connection
+        if passed is self._MARKER:
+            return _require_connection()
+        return _require_connection(passed)
+
+    def _monkey(self, connection):
+        from gcloud.datastore._testing import _monkey_defaults
+        return _monkey_defaults(connection=connection)
+
+    def test_implicit_unset(self):
+        with self._monkey(None):
+            with self.assertRaises(EnvironmentError):
+                self._callFUT()
+
+    def test_implicit_unset_w_existing_batch(self):
+        from gcloud.datastore._testing import _NoCommitBatch
+        ID = 'DATASET'
+        CONNECTION = object()
+        with self._monkey(None):
+            with _NoCommitBatch(dataset_id=ID, connection=CONNECTION):
+                self.assertEqual(self._callFUT(), CONNECTION)
+
+    def test_implicit_unset_w_existing_transaction(self):
+        from gcloud.datastore._testing import _NoCommitTransaction
+        ID = 'DATASET'
+        CONNECTION = object()
+        with self._monkey(None):
+            with _NoCommitTransaction(dataset_id=ID, connection=CONNECTION):
+                self.assertEqual(self._callFUT(), CONNECTION)
+
+    def test_implicit_unset_passed_explicitly(self):
+        CONNECTION = object()
+        with self._monkey(None):
+            self.assertTrue(self._callFUT(CONNECTION) is CONNECTION)
+
+    def test_implicit_set(self):
+        IMPLICIT_CONNECTION = object()
+        with self._monkey(IMPLICIT_CONNECTION):
+            self.assertTrue(self._callFUT() is IMPLICIT_CONNECTION)
+
+    def test_implicit_set_passed_explicitly(self):
+        IMPLICIT_CONNECTION = object()
+        CONNECTION = object()
+        with self._monkey(IMPLICIT_CONNECTION):
+            self.assertTrue(self._callFUT(CONNECTION) is CONNECTION)
+
+
 class Test_set_default_connection(unittest2.TestCase):
 
     def setUp(self):
