+++ gcloud/storage/_helpers.py
@@ -20,8 +20,7 @@ These are *not* part of the API.
 from Crypto.Hash import MD5
 import base64
 
-from gcloud.storage._implicit_environ import get_default_connection
-from gcloud.storage.connection import _CONNECTIONS
+from gcloud.storage._implicit_environ import _require_connection
 
 
 class _PropertyMixin(object):
@@ -111,29 +110,6 @@ class _PropertyMixin(object):
             method='PATCH', path=self.path, data=update_properties,
             query_params={'projection': 'full'}, _target_object=self)
         self._set_properties(api_response)
-
-
-def _require_connection(connection=None):
-    """Infer a connection from the environment, if not passed explicitly.
-
-    :type connection: :class:`gcloud.storage.connection.Connection`
-    :param connection: Optional.
-
-    :rtype: :class:`gcloud.storage.connection.Connection`
-    :returns: A connection based on the current environment.
-    :raises: :class:`EnvironmentError` if ``connection`` is ``None``, and
-             cannot be inferred from the environment.
-    """
-    if connection is None:
-        connection = _CONNECTIONS.top
-
-    if connection is None:
-        connection = get_default_connection()
-
-    if connection is None:
-        raise EnvironmentError('Connection could not be inferred.')
-
-    return connection
 
 
 def _scalar_property(fieldname):
