+++ gcloud/storage/test__helpers.py
@@ -155,44 +155,6 @@ class Test__scalar_property(unittest2.TestCase):
         self.assertEqual(test._patched, ('solfege', 'Latido'))
 
 
-class Test__require_connection(unittest2.TestCase):
-
-    def _callFUT(self, connection=None):
-        from gcloud.storage._helpers import _require_connection
-        return _require_connection(connection=connection)
-
-    def _monkey(self, connection):
-        from gcloud.storage._testing import _monkey_defaults
-        return _monkey_defaults(connection=connection)
-
-    def test_implicit_unset(self):
-        with self._monkey(None):
-            with self.assertRaises(EnvironmentError):
-                self._callFUT()
-
-    def test_implicit_unset_w_existing_batch(self):
-        CONNECTION = object()
-        with self._monkey(None):
-            with _NoCommitBatch(connection=CONNECTION):
-                self.assertEqual(self._callFUT(), CONNECTION)
-
-    def test_implicit_unset_passed_explicitly(self):
-        CONNECTION = object()
-        with self._monkey(None):
-            self.assertTrue(self._callFUT(CONNECTION) is CONNECTION)
-
-    def test_implicit_set(self):
-        IMPLICIT_CONNECTION = object()
-        with self._monkey(IMPLICIT_CONNECTION):
-            self.assertTrue(self._callFUT() is IMPLICIT_CONNECTION)
-
-    def test_implicit_set_passed_explicitly(self):
-        IMPLICIT_CONNECTION = object()
-        CONNECTION = object()
-        with self._monkey(IMPLICIT_CONNECTION):
-            self.assertTrue(self._callFUT(CONNECTION) is CONNECTION)
-
-
 class Test__base64_md5hash(unittest2.TestCase):
 
     def _callFUT(self, bytes_to_sign):
@@ -286,18 +248,3 @@ class _Base64(object):
     def b64encode(self, value):
         self._called_b64encode.append(value)
         return value
-
-
-class _NoCommitBatch(object):
-
-    def __init__(self, connection):
-        self._connection = connection
-
-    def __enter__(self):
-        from gcloud.storage.connection import _CONNECTIONS
-        _CONNECTIONS.push(self._connection)
-        return self._connection
-
-    def __exit__(self, *args):
-        from gcloud.storage.connection import _CONNECTIONS
-        _CONNECTIONS.pop()
