+++ gcloud/storage/test__implicit_environ.py
@@ -25,6 +25,44 @@ class Test_get_default_bucket(unittest2.TestCase):
         self.assertTrue(self._callFUT() is None)
 
 
+class Test__require_connection(unittest2.TestCase):
+
+    def _callFUT(self, connection=None):
+        from gcloud.storage._implicit_environ import _require_connection
+        return _require_connection(connection=connection)
+
+    def _monkey(self, connection):
+        from gcloud.storage._testing import _monkey_defaults
+        return _monkey_defaults(connection=connection)
+
+    def test_implicit_unset(self):
+        with self._monkey(None):
+            with self.assertRaises(EnvironmentError):
+                self._callFUT()
+
+    def test_implicit_unset_w_existing_batch(self):
+        CONNECTION = object()
+        with self._monkey(None):
+            with _NoCommitBatch(connection=CONNECTION):
+                self.assertEqual(self._callFUT(), CONNECTION)
+
+    def test_implicit_unset_passed_explicitly(self):
+        CONNECTION = object()
+        with self._monkey(None):
+            self.assertTrue(self._callFUT(CONNECTION) is CONNECTION)
+
+    def test_implicit_set(self):
+        IMPLICIT_CONNECTION = object()
+        with self._monkey(IMPLICIT_CONNECTION):
+            self.assertTrue(self._callFUT() is IMPLICIT_CONNECTION)
+
+    def test_implicit_set_passed_explicitly(self):
+        IMPLICIT_CONNECTION = object()
+        CONNECTION = object()
+        with self._monkey(IMPLICIT_CONNECTION):
+            self.assertTrue(self._callFUT(CONNECTION) is CONNECTION)
+
+
 class Test_get_default_connection(unittest2.TestCase):
 
     def setUp(self):
@@ -109,3 +147,18 @@ class Test_lazy_loading(unittest2.TestCase):
         self.assertTrue(isinstance(lazy_loaded, Connection))
         self.assertTrue(
             'connection' in _implicit_environ._DEFAULTS.__dict__)
+
+
+class _NoCommitBatch(object):
+
+    def __init__(self, connection):
+        self._connection = connection
+
+    def __enter__(self):
+        from gcloud.storage.connection import _CONNECTIONS
+        _CONNECTIONS.push(self._connection)
+        return self._connection
+
+    def __exit__(self, *args):
+        from gcloud.storage.connection import _CONNECTIONS
+        _CONNECTIONS.pop()
